dbsharding/requirements.txt
mysql-connector-python==9.0.0
PyYAML==6.0.3
